@@ -2,7 +2,7 @@
   "name": "edg/module-erp-koppeling",
   "description": "EDG ERP Magento 2 integration",
   "require": {
-    "php": "~5.6.0|~7.0.0|~7.1.0|~7.2.0|~7.3.0",
+    "php": "|~7.4.0",
     "edg/module-erp-service": "*",
     "magento/magento2-base": "^2.1.9"
   },
