@@ -3,23 +3,43 @@
 namespace Edg\Erp\Observer\Frontend;
 
 
+use Magento\Framework\App\Config\ScopeConfigInterface;
+use Magento\Framework\Event\Observer;
 use Magento\Framework\Event\ObserverInterface;
 use Magento\Framework\Exception\CouldNotSaveException;
+use Magento\Framework\Exception\LocalizedException;
+use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Quote\Api\Data\CartInterface;
+use Magento\Quote\Model\Quote;
 
 class PreventDuplicateSkuInQuote implements ObserverInterface
 {
-    protected $scopeConfig;
+    /**
+     * @var ScopeConfigInterface
+     */
+    protected ScopeConfigInterface $scopeConfig;
+
+    /**
+     * @var CartInterface|Quote
+     */
     protected $quote;
 
+    /**
+     * @throws NoSuchEntityException
+     * @throws LocalizedException
+     */
     public function __construct(
-        \Magento\Framework\App\Config\ScopeConfigInterface $configInterface,
+        ScopeConfigInterface $configInterface,
         \Magento\Checkout\Model\Session $session
     ) {
         $this->quote = $session->getQuote();
         $this->scopeConfig = $configInterface;
     }
 
-    public function execute(\Magento\Framework\Event\Observer $observer)
+    /**
+     * @throws CouldNotSaveException
+     */
+    public function execute(Observer $observer)
     {
         if (!$this->scopeConfig->isSetFlag('bold_orderexim/settings/block_duplicate_sku',
             \Magento\Store\Model\ScopeInterface::SCOPE_STORE)
